Allow unparse_url() to fetch /security.txt instead
A second, optional parameter has been added to unparse_url() to allow it
to fetch a security.txt file from the legacy path, /security.txt, as
opposed to the current path, /.well-known/security.txt.

This currently does nothing, but a future version of the validator may
support fetching from the deprecated location.

Ideally, webmasters should be redirecting /security.txt to
/.well-known/security.txt.